move types to their own package
