+++ automerge-js/test/basic_test.ts
@@ -1,3 +1,4 @@
+import * as tt from "automerge-types"
 import * as assert from 'assert'
 import * as util from 'util'
 import * as Automerge from '../src'
