+++ automerge-wasm/examples/webpack/package.json
@@ -10,7 +10,7 @@
   },
   "author": "",
   "dependencies": {
-    "automerge-wasm": "file:automerge-wasm-0.1.3.tgz"
+    "automerge-wasm": "file:automerge-wasm-0.1.4.tgz"
   },
   "devDependencies": {
     "serve": "^13.0.2",
