+++ automerge-wasm/package.json
@@ -8,7 +8,7 @@
   "description": "wasm-bindgen bindings to the automerge rust implementation",
   "homepage": "https://github.com/automerge/automerge-rs/tree/main/automerge-wasm",
   "repository": "github:automerge/automerge-rs",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "license": "MIT",
   "files": [
     "README.md",
@@ -38,6 +38,7 @@
     "@types/jest": "^27.4.0",
     "@types/mocha": "^9.1.0",
     "@types/node": "^17.0.13",
+    "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.25.0",
     "@typescript-eslint/parser": "^5.25.0",
     "cross-env": "^7.0.3",
@@ -48,5 +49,8 @@
     "rimraf": "^3.0.2",
     "ts-mocha": "^9.0.2",
     "typescript": "^4.6.4"
+  },
+  "dependencies": {
+    "automerge-types": "^0.1.1"
   }
 }
